Update util.py
This helps with #26. Needed a way to deal with unexpected errors that come up. I couldn't figure out a better way so maybe someone else could. This error wasn't just tied to django-honeypot. For example, making a session variable and then deleting it from the database caused the same type of problem.