:recycle: [#13] -- refactor clean cookies middleware implementation
This should make it easier to understand how it works
and what it's supposed to be doing.

Note: the case where cookie version is compared with group version
is removed. It's unclear what the logic behind this is, and it's
probably also wrong because the group version is derived from the
first item cookie in the cookie group, but this has no explicit
ordering and may be based on database ID or timestamp, so
it's at least unreliable on different database backends.