:package: Build JS assets in CI and gitignore them
The npm 'package' and the django staticfiles assets are exclusively
to be built in the CI pipeline, and not kept in version control.

This keeps a single source of truth and prevents us from forgetting to
update the artifacts, at the cost of a little extra step for local
development.