:sparkles: Fixes #103 -- support natural keys for the models
Added support for natural keys for the cookie and cookie group
models. This required adding another unique constraint, this
time to the cookie model. A specific cookie within a group was
already looked up by name and domain in the existing code, so
this DB constraint confirms the usage. Duplicate data would not
have yielded the expected results, but users may not be aware
of it, so this is also a breaking change.

You can now dump and load fixtures for cookie consent, using
these natural keys and avoid database primary key conflicts.