:thread: Fix race condition in test
Clicking the accept button fires a fetch call, which sometimes
completes before playwright navigates to the page again,
sometimes not.

This results in the accepted cookies state not being recorded,
in turn causing the test to fail.

Adding an assertion guarantees that the fetch call completed
before we 'refresh' the page.