:rotating_light: Enable the pyupgrade rules
Automatic upgrade for the modern type annotations. Some manual
fixes were required for the string interpolation.

Along the way, the varname for an individual cookie is deprecated.

A code search in Github revealed that the accepted_cookies template
tag does not appear to be used in any public projects. Deprecating
it should be safe - I also can't really imagine how/why you'd use
this tag as it's documented and more common to check whether an
entire cookie group is accepted.